refactor: move LoginResponse type and validation to utils for better organization
